scroll to top as component, management moved to constants
