+++ src/components/constants.js
@@ -1,0 +1,28 @@
+const management = [
+  "Aidan Hollis",
+  "Peggy Tse",
+  "Maria Ginevra Cattaneo",
+  "Max Alexander Matthey",
+  "Felicitas Holzer",
+  "Daniele Botti",
+  // "Najid Ahmad",
+  // "Özlem Denli",
+  // "Seung Chan (Kevin) Lee",
+  "Boris Yakubchik",
+];
+
+// must coincide with order of `management`
+const managementImages = [
+  "Aidan.jpg",
+  "Peggy.jpg",
+  "Maria.jpg",
+  "Max.jpg",
+  "Felicitas.jpg",
+  "Daniele.jpg",
+  // "",
+  // "Ozlem.jpg",
+  // "",
+  "Boris.jpg",
+];
+
+module.exports = { management, managementImages };
