+++ src/pages/how.js
@@ -3,6 +3,7 @@ import { FormattedMessage, injectIntl } from "gatsby-plugin-intl"
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
+import ScrollToTop from "../components/scrolltotop"
 
 const Benefits = ({ intl }) => (
   <Layout>
@@ -12,26 +13,7 @@ const Benefits = ({ intl }) => (
       description={intl.formatMessage({ id: "general.description" })}
     />
 
-    <div
-      className="scroll-to-top"
-      role="button"
-      onClick={() => {
-        window.scrollTo(0, 0)
-      }}
-    >
-      <svg
-        width="16"
-        height="10"
-        viewBox="0 0 16 10"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M14.1053 0L8 6.10526L1.89474 0L0 1.89474L8 9.89474L16 1.89474L14.1053 0Z"
-          fill="#222222"
-        />
-      </svg>
-    </div>
+    <ScrollToTop />
 
     <h1 id="summary">
       <FormattedMessage id="how.in_short" />
